more quad tree memes
