+++ test/src/Main.hs
@@ -9,14 +9,16 @@ module Main (main) where
 
 import Test.QuadTree
 import Test.Sandwich
+import Test.Sandwich.QuickCheck
 
 basic :: TopSpec
 basic = do
-  describe "Quad tree" $ do
-    it "inserts into a tree" $ do
-      (2 + 2) `shouldBe` 4
-    it "Makes sure that the bounds check works correctly" $ do
+  describe "Quad tree tests" $ do
+    it
+      "Ensures that the bounds check works correctly"
       testBounds
+    introduceQuickCheck $ do
+      prop "inserting out of bounds elements doesn't change the quad tree" propInsertOutOfBounds
 
 main :: IO ()
 main = do
