+++ test/src/Test/QuadTree.hs
@@ -1,7 +1,3 @@
-{-# LANGUAGE MultiParamTypeClasses #-}
-{-# LANGUAGE ScopedTypeVariables #-}
-{-# OPTIONS_GHC -Wno-orphans #-}
-
 -- |
 --    Module      : Test.QuadTree
 --    License     : GNU GPL, version 3 or above
@@ -13,85 +9,35 @@ module Test.QuadTree
   ( Position (..)
   , testBounds
   , testTree
+  , propInsertOutOfBounds
   )
 where
 
-import Control.Applicative
 import Control.Exception (SomeException)
 import Control.Monad.Catch (MonadCatch, try)
 import Data.Foldable (traverse_)
-import GHC.Generics (Generic)
 import GHC.Stack.Types (HasCallStack)
 import Myopia.QuadTree
 import Myopia.QuadTree.Internal
-import Myopia.Util
 import Test.QuickCheck
 import Test.Sandwich
-
--- Dummy data type for inserting into quad trees
-newtype Position = Pos (Double, Double)
-  deriving stock (Show, Generic)
-
--- Must implement the HasPos interface for quad tree to work
-instance HasPos Position Double where
-  getPosition (Pos (x, y)) = P (V2 x y)
-
-instance Arbitrary Position where
-  arbitrary = liftA2 (Pos ... (,)) arbitrary arbitrary
-
-instance Arbitrary a => Arbitrary (Quadrant a) where
-  arbitrary = sized tree
-    where
-      tree 0 = Leaf <$> arbitrary
-      tree n =
-        oneof
-          [ tree 0
-          , Node <$> subtree <*> subtree <*> subtree <*> subtree
-          ]
-        where
-          subtree = tree (n `div` 4)
-
-instance Arbitrary a => Arbitrary (V2 a) where
-  arbitrary = V2 <$> arbitrary <*> arbitrary
-
-instance (Num i, Ord i, Arbitrary i) => Arbitrary (Boundry i) where
-  arbitrary = Boundry <$> location <*> width <*> height
-    where
-      location = P ... V2 <$> arbitrary <*> arbitrary
-      width = getPositive . Positive <$> arbitrary
-      height = getPositive . Positive <$> arbitrary
-
-hasSize :: Int -> Quadrant [a] -> Bool
-hasSize n (Leaf xs) = n >= length xs
-hasSize n (Node nw ne sw se) = all (hasSize n) [nw, ne, sw, se]
-
-instance (Num i, Ord i, Arbitrary i, Arbitrary a) => Arbitrary (QuadTree i a) where
-  arbitrary = do
-    capacity <- arbitrary
-    region <- arbitrary `suchThat` hasSize capacity
-    boundry <- arbitrary
-    pure $ QuadTree region boundry capacity
+import Test.Types
 
 -- TODO: implement property checks
---
--- inserting out of bounds point doesn't change the quad tree
--- insert a qt = qt
 --
 -- inserting and then querying all in bounds should return the same elements
 -- elemsInBoundry (insertElems inbounds qt) = inbounds
 --
--- if an element is inbounds then dividing the boundry should only should
+-- if an element is inbounds then dividing the boundry should
 -- only cause that element to be inbounds of 1 out of the 4 new boundries
 -- inBounds elem boundry = True -> length (filter inBounds (divide boundry)) = 1
 
--- propInsertOutOfBounds :: Gen Property
--- propInsertOutOfBounds = do
---   randQuadTree <- arbitrary @(QuadTree Int (Boundry Int))
---   bound <- arbitrary @(Boundry Int)
---   let outOfBounds = not . inBounds bound
---       quadTree = randQuadTree {boundry = bound}
---   position <- arbitrary `suchThat` outOfBounds
---   pure $ insert position quadTree === quadTree
+propInsertOutOfBounds :: Gen Property
+propInsertOutOfBounds = do
+  quadTree <- arbitrary @(QuadTree Double Position)
+  let outOfBounds = not . flip inBounds quadTree.boundry
+  position <- arbitrary `suchThat` outOfBounds
+  pure $ insert position quadTree === quadTree
 
 -- Test that the inBounds function behaves correctly
 testBounds :: (HasCallStack, MonadCatch m) => m ()
@@ -102,7 +48,7 @@ testBounds =
   where
     boundry =
       Boundry
-        { location = P (V2 100 100)
+        { center = P (V2 100 100)
         , width = 60
         , height = 50
         }
